Move adapters into own namespace
